Add support for setting TX power on NanoVNA v2
As of firmware version 1.0.2 (tag 20210214) the output power of the TX
ADF4350 can be configured via USB (register 0x42). Add support for
this setting in NanoVNASaver.

The SI5351 does not support setting the output voltage / power (only
the maximum current which controls rise / fall time rather than output
power) so we can only affect frequencies from 140MHz up. Clearly show
this in the UI so there are no surprises.

Because we can only detect this capability by talking to the VNA, the
setting is populated during connect.

*WARNING*: The automatic first sweep after connecting to the VNA will
be performed at maximum power (default). Sensitive equipment should
not be connect until after NanoVNASaver has connected to the NanoVNA
v2 and the setting has been changed.