Cast progress value to integer
PyQt5 has changed to hint the progress value to integer, resulting in a
type mismatch and subsequent crash when we're passing a float in. This
is often the case when using more than one segment. Rather than crash,
we should just truncate the progress value to integer.

I've chosen truncation rather than rounding to ensure values like 99.5% are
pessimistically rounded down rather than up, avoiding the false
impression that all sweeps are done.