__main__.py: remove /bin/env shebang
The file is not intended for direct execution,
not installed in the path,
and not marked as executable.
Its extension is sufficient for editors to trigger syntax highlighting.

The shebang seems to only trigger warnings on systems wher /bin/env is
unavailable (it may be in /usr/bin for example).