fix: sweep settings and hw version update (#628)
* Remove privacy breaches from the HTML documentation

Icons and screenshots should not inform distant websites each time an
user reads readme.html.

The suggestions in this draft are independent and will probably be
discussed separately, but they affect a single file so for a first
review a single commit is convenient.

* fix: sweep settings and hw version update

NamedTuples are inmutable either use _replace or use the class itself.

* Revert "Remove privacy breaches from the HTML documentation"

---------

Co-authored-by: Nicolas Boulenguez <nicolas@debian.org>