Fix checking for updates (#674)
* Because the application version seems to be a generated value, a different
method needed to be devised to try and determine what the current version
nuber if. We now attempt scan the github release tags for the latest release
version. If a newer version is detected, the application will display the
current version and direct the user to click the link to view the latest
release page.

* Rather, should report error against TAGS_URL and not LATEST_URL at this stage.